Reorganize sign up folder
